# docs/swagger.yaml
consumes:
- application/json
definitions:
...
  ErrorResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      message:
        type: string
      path:
        type: string
      status:
        type: integer
      timestamp:
        type: string
    type: object
  PatientPatch:
    properties:
      address:
        type: string
      admission_date:
        type: string
      dni:
        type: string
      email:
        type: string
      last_name:
        type: string
      name:
        type: string
    type: object
  PatientPost:
    properties:
      address:
        type: string
      admission_date:
        type: string
      dni:
        type: string
      email:
        type: string
      last_name:
        type: string
      name:
        type: string
    required:
    - address
    - admission_date
    - dni
    - email
    - last_name
    - name
    type: object
  PatientPut:
    properties:
      address:
        type: string
      admission_date:
        type: string
      dni:
        type: string
      email:
        type: string
      last_name:
        type: string
      name:
        type: string
    required:
    - address
    - admission_date
    - dni
    - email
    - last_name
    - name
    type: object
  PatientResponse:
    properties:
      address:
        type: string
      admission_date:
        type: string
      dni:
        type: string
      email:
        type: string
      id:
        type: integer
      last_name:
        type: string
      name:
        type: string
    type: object
...
info:
  contact:
    email: gabriela.calicanton@gmail.com
    name: Gabriela Cecilia Calicanton
    url: http://www.swagger.io/support
  description: API of Dental Clinic
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Dental Clinic API
  version: "2.0"
paths:
...
  /patients:
    get:
      description: Get all Patients
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/PatientResponse'
            type: array
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get all Patients
      tags:
      - Patient
    post:
      description: Create a Patient
      parameters:
      - description: Public Key
        in: header
        name: PUB_KEY
        required: true
        type: string
      - description: PatientResponse
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/PatientPost'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/PatientResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - APIKey: []
      summary: Create a Patient
      tags:
      - Patient
  /patients/{id}:
    delete:
      description: Delete a Patient
      parameters:
      - description: Public Key
        in: header
        name: PUB_KEY
        required: true
        type: string
      - description: Patient ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - APIKey: []
      summary: Delete a Patient
      tags:
      - Patient
    get:
      description: Get Patient by id
      parameters:
      - description: Patient ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PatientResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get Patient by id
      tags:
      - Patient
    patch:
      description: Patch a Patient
      parameters:
      - description: Public Key
        in: header
        name: PUB_KEY
        required: true
        type: string
      - description: Patient ID
        in: path
        name: id
        required: true
        type: integer
      - description: PatientResponse
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/PatientPatch'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PatientResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - APIKey: []
      summary: Patch a Patient
      tags:
      - Patient
    put:
      description: Update a Patient
      parameters:
      - description: Public Key
        in: header
        name: PUB_KEY
        required: true
        type: string
      - description: Patient ID
        in: path
        name: id
        required: true
        type: integer
      - description: PatientResponse
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/PatientPut'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PatientResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - APIKey: []
      summary: Update a Patient
      tags:
      - Patient
  /patients/q:
    get:
      description: Get Patient by DNI
      parameters:
      - description: Patient DNI
        in: query
        name: dni
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PatientResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get Patient by DNI
      tags:
      - Patient
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  APIKey:
    description: Add secret key here
    in: header
    name: SECRET_KEY
    type: apiKey
swagger: "2.0"
...